UNHCR preferred graphic colors to be use as a main set for graphics creation.

pal_unhcr_graphic provides an additional UNHCR color palette that collects the preferred UNHCR colors for graphics as a data frame containing 6 colors.

all_palunhcr_graphic

Format

An object of class character of length 7.

Details

See https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html for details.

See also

Examples

pal_unhcr_graphic

#>         Blue Darkblue   Green   Grey3     Red   Blue3
#> Blue #0072BC  #18375F #00B398 #999999 #EF4A60 #66AAD7

dim(pal_unhcr_graphic)  # 1 6

#> [1] 1 6

# Access by position:
pal_unhcr_graphic[1]    # color blue (as df)

#>         Blue
#> Blue #0072BC

pal_unhcr_graphic[[1]]  # color value "#0072BC"

#> [1] "#0072BC"

# Access by name:
pal_unhcr_graphic["blue"]   # color "blue5" (as df)

#> Error in `[.data.frame`(pal_unhcr_graphic, "blue"): undefined columns selected

pal_unhcr_graphic[["blue"]] # color value "#0072BC"

#> NULL

# Plotting palette:
seecol(pal_unhcr_graphic)

#> Error in lapply(all_palun, exists): object 'all_palun' not found