Custom color and fill scales Custom coloring and filling functions based on unique color palettes Specific functions include:

scale_(color/colour/fill)_discrete_unhcrDiscrete palette with either fixed or dynamically extended number of shades

scale_(color/colour/fill)_opinionated_unhcrDiscrete palette with specific values for "good", "bad", and "neutral"

scale_(color/colour/fill)_diverging_unhcrContinuous diverging color palette

scale_(color/colour/fill)_continuous_unhcrCOntinuous color palette

Custom color and fill scales

Custom coloring and filling functions based on unique color palettes

Specific functions include:

scale_(color/colour/fill)_discrete_unhcrDiscrete palette with either fixed or dynamically extended number of shades

scale_(color/colour/fill)_opinionated_unhcrDiscrete palette with specific values for "good", "bad", and "neutral"

scale_(color/colour/fill)_diverging_unhcrContinuous diverging color palette

scale_(color/colour/fill)_continuous_unhcrCOntinuous color palette

Discrete palette with either fixed or dynamically extended number of shades

Discrete palette with specific values for "good", "bad", and "neutral"

Continuous diverging color palette

COntinuous color palette

scale_color_discrete_unhcr(palette = "test", extend = FALSE, ...)

scale_colour_discrete_unhcr(palette = "test", extend = FALSE, ...)

scale_fill_discrete_unhcr(palette = "test", extend = FALSE, ...)

scale_color_opinionated_unhcr(palette = "test", ...)

scale_colour_opinionated_unhcr(palette = "test", extend = FALSE, ...)

scale_fill_opinionated_unhcr(palette = "test", ...)

scale_color_diverging_unhcr(palette = "test", ...)

scale_colour_diverging_unhcr(palette = "test", ...)

scale_fill_diverging_unhcr(palette = "test", ...)

scale_color_continuous_unhcr(palette = "test", ...)

scale_colour_continuous_unhcr(palette = "test", ...)

scale_fill_continuous_unhcr(palette = "test", ...)

Arguments

palette
Name of color palette

extend
Whether to extend discrete color palette to make sufficient colors for levels needed

...
Additional arguments to be passed to internal scale function

References

https://en.wikipedia.org/wiki/Traffic_light